requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-core==1.4.1
google-cloud-pubsub==2.1.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.12
MarkupSafe==1.1.1
multidict==4.7.6
mypy-extensions==0.4.3
proto-plus==1.10.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.0
